added ArmstrongNumber.js
When the user enters a number, it is taken as a string. The length property returns the length of a string.
The number entered by the user is stored in a temp variable. And a while loop is used to iterate until its value is greater than 0. Each digit of the number is raised to the power of the length of the number.