quick_sort.cpp
In Insertion Sort, you simply went through each element in order and inserted it into a sorted sub-array. In this programm, you cannot focus on one element at a time, but instead must deal with whole sub-arrays, with a strategy known as "divide and conquer".When partition is called on an array, two parts of the array get 'sorted' with respect to each other. If partition is then called on each sub-array, the array will now be split into four parts. This process can be repeated until the sub-arrays are small. 
In this programm, I print array every time partitioning method finishes, i.e. whenever two subarrays, along with the pivot, are merged together